Refactor banner component and add lower banner component
The banner component was removed and replaced with a new lower banner component. This change improves code structuring and enhances the home page layout.
